Changes requested positive output, more rspec test
