+++ Gemfile
@@ -6,3 +6,4 @@ git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
 gem 'httparty'
 gem 'nokogiri'
+gem 'twitter'
